pyproject.toml
[tool.poetry]
name = "mytapo"
version = "0.1.0"
description = ""
authors = ["Fabian Wörenkämper <f.woerenkaemper@krombacher.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
asyncio = "^3.4.3"
matplotlib = "^3.9.4"
tapo = "^0.8.0"
ipython = "^8.30.0"
ipykernel = "^6.29.5"
requests = "^2.32.3"
influxdb-client = "^1.48.0"
notebook = "^7.3.2"
pyarrow = "^19.0.1"
fastparquet = "^2024.11.0"
watchdog = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
